+++ sacrerouge/tests/metrics/decomposed_rouge/decomposed_rouge_test.py
@@ -1,3 +1,5 @@
+import spacy
+
 from sacrerouge.common.testing import FIXTURES_ROOT
 from sacrerouge.common.testing.metric_test_cases import ReferenceBasedMetricTestCase
 from sacrerouge.common.testing.util import sacrerouge_command_exists
@@ -6,6 +8,11 @@ from sacrerouge.metrics import DecomposedRouge
 
 
 class TestDecomposedRouge(ReferenceBasedMetricTestCase):
+    def test_spacy_model_version(self):
+        # These results rely on having spacy en_core_web_sm version 2.3.1 installed
+        nlp = spacy.load('en_core_web_sm')
+        assert nlp.meta['version'] == '2.3.1'
+
     def test_regression(self):
         # This is a regression test, not necessarily a test for correctness. The expected output
         # is really long, so we saved it in a file