+++ CHANGELOG.md
@@ -7,3 +7,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added ability to skip calculating specific correlation levels (summary, system, and global)
+- Updating `spacy` package version to `2.3.3` and model version to `2.3.1`.
+`DecomposedRouge`'s unit tests and experiments subsequently updated.